Refactor `user` into forked_user and forked_group
Refactor the single `user` attribute into two separate attributes (one
for user and one for group). This makes it more clear as to how you
would specify a group for forked child processes. Also add loggging to
alert the end user if they only specified a group and not a user.
Unicorn does not support dropping group privileges only and requires
that a user also be specified.